Make all images work without logging warnings
