+++ src/components/PassthroughImage.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import { StaticQuery, graphql } from 'gatsby';
+
+const PassthroughImage = ({ filename, alt='' }) => (
+  <StaticQuery
+    query={graphql`
+      query {
+        images: allFile(filter: {extension: {regex: "/(gif)|(svg)/"}}) {
+          edges {
+            node {
+              relativePath
+              publicURL
+            }
+          }
+        }
+      }
+    `}
+    render={data => {
+      const image = data.images.edges.find(n => {
+        return n.node.relativePath.includes(filename);
+      });
+
+      if (!image) {
+        return null;
+      }
+
+      return <img src={image.node.publicURL} alt={alt} />;
+    }}
+  />
+); 
+
+export default PassthroughImage;
