+++ src/mdx-pages/blog/introducing-video-js-http-streaming-vhs.mdx
@@ -25,7 +25,7 @@ This was the motivation behind integrating [Video.js HTTP Streaming](https://git
 
 ## What is VHS?
 
-<BlogImage filename="blog/introducing-video-js-http-streaming-vhs/vhs-logo.svg" />
+<BlogImage filename="blog/introducing-video-js-http-streaming-vhs/vhs-logo.svg" alt="VHS logo"/>
 
 VHS is the successor to videojs-contrib-hls. It is a source handler forked from the [videojs-contrib-hls](https://github.com/videojs/videojs-contrib-hls) repository. While videojs-contrib-hls was originally designed to add HLS playback on all browsers, we realized that the engine could also play other formats just as well.
