+++ src/mdx-pages/blog/video-js-7-is-here.mdx
@@ -9,7 +9,7 @@ tags:
 
 import BlogImage from '../../components/BlogComponents/BlogImage'
 
-<BlogImage filename="blog/video-js-7-is-here/vjs 7 small.png" presentationWidth={141} />
+<BlogImage filename="blog/video-js-7-is-here/vjs 7 small.png" presentationWidth={141} alt="" />
 
 I'm happy to announce that Video.js 7 is now out! It brings to Video.js support for HLS and experimental DASH support via the [Video.js HTTP Streaming (VHS)](https://github.com/videojs/http-streaming) project. Video.js 7 also drops support for some older Internet Explorer browsers only keeping support for IE11.
