+++ src/mdx-pages/blog/videojs-contrib-ads-6.mdx
@@ -29,7 +29,7 @@ In version 6, the state machine was refactored to match this strict definition o
 
 Here is a diagram of the new states. Blue states are content states and yellow states are ad states:
 
-<BlogImage filename="blog/videojs-contrib-ads-6/ad-states.png" />
+<BlogImage filename="blog/videojs-contrib-ads-6/ad-states.png" alt="Diagram of ad states" />
 
 For a history of how the state machine has evolved over time, [there is detailed information in this Github issue](https://github.com/videojs/videojs-contrib-ads/issues/320). In this article we're going to focus on how it works in version 6.
